Add standard Copy + Select All context menu to the SettingsTextBox.
